pyproject.toml
[project]
name = "gigabyte-rag-assistant"
version = "0.1.0"
description = "純 Python RAG 問答系統，針對 GIGABYTE AORUS MASTER 16 AM6H 產品規格，支援繁體中文與英文混合查詢"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.89.0",
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "httpx>=0.28.1",
    "llama-cpp-python>=0.3.19",
    "numpy>=2.4.4",
    "playwright>=1.58.0",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "requests>=2.33.1",
    "rich>=14.3.3",
    "sentence-transformers>=5.3.0",
]
